requirements.txt
Flask>=2.2
Flask-SocketIO>=5.3.2
eventlet>=0.33.0
gunicorn>=20.1.0
python-docx>=0.8.11
PyMuPDF>=1.22.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0
openpyxl>=3.1.2
phonenumbers>=8.13.11
openai>=1.50.0
python-dotenv>=1.0.0
tqdm>=4.65.0
celery[redis]>=5.2.3
redis>=4.5.0